+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## next
+\* Improve documentation.<br/>
+
 ## 0.1.0
 \+ Add TypedJSON class.<br/>