+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 \+ Add type information to `.is` checks.<br/>
 \+ Add `.keys()`.<br/>
+\+ Add `.toString()`.<br/>
 \* Replaced computed properties with methods.<br/>
 \* Improve documentation.<br/>